requirements.txt
certifi==2019.3.9
chardet==3.0.4
clearbit==0.1.7
Django==2.2
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
idna==2.8
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.0
urllib3==1.24.1
